added new folders to organize code better
