+++ client/app.jsx
@@ -4,7 +4,7 @@ import GuestList from './pages/guest-list';
 
 export default class App extends React.Component {
   render() {
-    // return <InspirationBoard />;
     return <GuestList />;
+    // return <InspirationBoard />;
   }
 }
